requirements.txt
# Trading Companion 2026 Dependencies
# Install with: pip install -r requirements.txt

# Core
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0

# API clients
alpaca-py>=0.13.0

# CLI
rich>=13.0.0  # Optional: prettier terminal output

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
responses>=0.23.0  # For mocking HTTP

# Development
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
